[feat] Adding connexion from API to DB. Creating /send_data API endpoint.
US #11

Creating /send_data endpoint used to store a new data entry inside the Data table in the database. Currently using a
hardcoded date to store in the data entry.

Notable problem encountered : When connecting to the mariadb container from the api container I used host "localhost".
The api container was looking for a database on itself and not the localhost (where the containers run.).