[feat] Adding transformDate function to "align" dates in the same timeframe
US #11

When powering back on the raspberry it's internal clock will restart at a certain date (1 Jan. 2021, 1 Jan. 2015, 1 Jan.
1970, ... -> need to check). The problem will be to knwo exactly when the measures were taken.

To do that we send the date of the measure and the date at which the data is sent (will be usefull later when sending
multiple measured data). On the api's side we will retrieve the date the request in sent and modify the measure date
compared to the actal date. This way we will have the correct dates at which the data was measured.