[feat] Completing Chart component for it to call /monitoring_data endpoint
US #18

Adding the call to the /monitoring_data endpoint to display its data in the chart. We can't use the mariadb container in
the axios url endpoint call beceause the call is not done from the frontend container but from the user's browser. We
will have to change that IP address to the deployement url later.
 # Please enter the commit message for your changes.