[feat] Changing CheckPasswordMatch so it checks if the account is available
US #2

If we call the ChechPasswordMath procedure with an email that does not exist, we will have a problem.

Adding a select to count if the email is available inside the database. If it does we check the password as before. If not, the
procedure returns a message indicating so.