[fix] Instead of refusing a connection when the session already exists, deletes it and creates a new one.
US #2

If the user deletes it's session id cookie and tries to reconnect it will fail beceause there already is a session in the DB. The user
would have to wait for the session to expire in the DB.

To avoid this we don't refuse a new connection if there still is an active session in the DB. We delete it and create a new one. A new
cookie is sent to the user.