[fix] Setting up connection endpoint and get_monitoring endpoint for brwser usage
US #2

Endpoints worked on Postman and set cookies and used them properly. It didn't work on a browser. With withCredentials to true when
calling the /user/connection endpoint there was a problem. I thought the cookie was not received and not set. Using the "hhtpOnly"
parameter when creating a cookie makes it unaccessible from frontend. It is only send during http communication (ex: call another
endpoint).

Setting "secure" parameter to true only allow its use with https or on localhost.

The roblem was I used withCredentials when using /user/connection endpoint but not when using /data/get_monitoring. For the first
endpoint it received the signed cookie and set it. But it wasn't sent for /data/get_monitoring so it couldn't be used by the endpoint
to work correctly. I thought I had it all working with "secure : true" but apparently it doesn't.