[refactor] /user/connect -> user already connected doesn't recreate a session in DB
US #2

Before, if the user tried to connect while he still had a session (with cookie)or if he had no cookie anymore, a new session was
created in the DB and a new cookie was given to him.

Now we first check if he has a valid session (with cookie and session in db). If yes, nothing is done. If not he is connected again.
If there is still an active session in the DB it is reused and a cookie with it's session id is given to the frontend.