[feat] Creating store_data() function on Raspberry Pi Pico
US #13

To avoid loosing data if the connexion from a probe to the database is lost. Before sending data to the DB we store it
locally on the probe.

The store_data() function stores data inside a data.json file on the raspberry. Each set of data is written inside that file
and is identified by the timestamp at which the data was taken.