[refactor] Timestamps in data.json are strings instead of arrays
US #13

If we use timestamps as arrays as attributes inside a JSON we will have problems to access it. So we use strings instead. We
only concatened the arrays elements with a blank in between each one. Will have to retransform into array on the API side
when sending data.
 # Please enter the commit message for your changes. Lines starting