[feat] /data/send_data can send multiple measures at once
US #22

If we store data on the probe we want to be able to send it later when the DB is accessible. We need to be able to send
multiple measurements at once.

Still sending probe_id and send_timestamp but also sending multiple dataset. Each dataset is represented by a timestamp as
attribute and it's temperature, float density and refraction density as a json.