[feat] Adding delete_data_older_than() function on probe
US #23

After locally stored data are send to the DB we need to delete the data that has been registered.

The delete_data_older_than function receives a timestamp as parameter and deletes every data inside data.json that is older
than the given timestamp. All other data is kept.