[refactor] Changing delete_data_older_than() to delete_oldest_records()
US #23

The base idea was to delete records in data.json with the timestamp of the oldest data stored in DB. But beceause we modify
the timestamps when sending the data the timestamps are not the same anymore.

Instead, deleting a certain number of the oldest records. The number of records to be deleted is given by a response of the
/data/send_data endpoint. It is the number of records that were stored in DB.