[feat] Changing sending_data() function on probe to use stored data from data.json
US #24

When storing data locally on the probe we need to send them later when the connection to the DB is restored.

Now, sending_data doesn't receive the data to send directly from parameters but fetch them from data.json. It still has
probe_id and send_timestamp as parameters. Before sending data, the new data that was measured is stored in data.json