[feat] Adding CheckProbePassword SQL procedure
US #12

When sending data from a probe we need to check the user does posess the probe. Therefor we use a password related to the
probe.

CheckProbePassword SQL procedure takes two parameters : probe_id and password. When the password is correct, the procedure
returns "correct as Response. If not, returns "not correct".