[fix] Changinf Dockerfiles and docker_compose.yaml volumes due to "Error loading shared library"
Originally we copied the package.json and package-lock.json of api and frontend. When running npm install in the Dockerfiles
we create node_modules directories. However, when using bind mounts (volumes in docker-compose.yaml) we change the
node_modules tree causing problems for certain packages.

Not using volumes and copying all files from host to container in the Dockerfiles fixes the issue. Also adding a
.dockerignore file with /node_modules entry to stop the node_modules directory from being copied inside the container and
replacing the tree (we would have the exact same problem).
starting # with '#' will be ignored, and an empty message aborts the commit.