[feat] Adding hashPassworWithArgon2 and checkPasswordArgon2 functions inside api_functions.js
US #12

hashPasswordWithArgon2 is a function that takes a password and argon2.hash() hashes it with a random salt. It will be used during
user and probe registration to store the password and the salt inside the DB. Originally we wanted to store those in two different
columns of our user and probe table. It is easier to store them in one columns as a string.

checkPasswordArgon2 will later use the stored hashed password (in DB) and compare it to a clear password. By using the parameters
inside the
hashed password it will calculate the salt and rehash the password. It will then compare the result with the stored hashed
password.
 # Please enter the commit message for your changes. Lines starting