[refactor] Changing tables and insert script to delete salt column in User and Probe tables and insert a hashed password using Argon2
Originally we wanted to use a column for the password and a column for the password and a column for the salt inside User and
Probe tables. However it is easier to store as one string when using Argon2 for backend hashing. When hashing a passwor Argon2
generates a string with multipe parameters and the hashed password. To compare a given password to the stored hash Argon2 will
recalculate the salt using the parameters of the stored string in DB. It will then rehash the password and compare the two hashes.