[feat] Replacing CheckProbePassword SQL procedure with RetrieveProbePassword
US #12

With the use of argon2 to hash and compare passwords we won't compare password on the DB.

RetrieveProbePassword will retireve the hashed password of a given probe. The comparisson will be done on the API where we use
argon2.verify() to compare the stored hash and the given password.