[fix] fixing InsertData SQL procedure, it was selecting the monitoringId in a bad way.
US #22

The procedure was selecting the monitorId in which we want to insert the data with only the probe id as parameter of the
subquery. So it was selecting all monitorings, including the finished ones. It couldn't store it inside an int :
"Subquery returns more than 1 row".

Adding "AND EndDate IS NULL" int the subquery's WHERE. Also removed condition where
it selects multiple monitoring ids beceause there
shouldn't be multiple active monitorings.
 # Please enter the commit message for your changes. Lines starting