[fix] Checking body request for endpoint /user/connect
US #20

Was not checking values of the body request for /user/connect. Plus I was not checking if the account existed or if the
password was correct. With a bad password or user email the endpoint just crashed.

Now doing all that. Checking if the email is a string and looks like a string. Checking if the passwor is a string and is
non_empty.