[feat] Storing isConnected state into the locale storage to allow the website to display certain components when connected or not.
We need to display certain components when the user is connected and others when he is not. For example we want a
disconnect button only if the user is connected and an access to the connection form only if he is not already connected.

UserContext.js contains the way we store and update isConnected. isConnected is stored in loaclStorage and is updated
with setIsConnected. We use UserProvider as a wrapper on the most upper layer of the components tree. This way all
children components can access it when using useContext.
 # Please enter the commit message for your changes. Lines