[feat] Adding /monitoring/get_archived api endpoint to retrieve all archived monitorings of a connected user
US #17

The endpoint checks if the user is connected and retrieves his user id thanks to his session cookie. We then call the
GetArchivedMonitoringsByUser sql procedure with the user id as parameter. We get back all the monitoring ids of the
archived monitorings. # Please enter the commit message for your changes. Lines starting