[feat] Adding ArchivedList react component
US #17

A user needs to be able to see the archived monitorings (and access their data).

This ArchivedList react compoenent uses the /monitoring/get_archive and /monitoring/get_data api endpoints to dynamically
generate a table with all the archived monitorings related to a user. Clicking on one of them directs to the Chart
component of the monitoring. # Please enter the commit message for your changes. Lines starting