[refactor] Changing docker volumes and copy in dockerfiles
Before, had a problem with argon2 library beceause we copied the package.json and npm installed all packages. When using the volumes in docker
compose the /node_modules tree was modified, causing a problem. Had deleted the volumes and copied every file in dockerfiles. But I had to rebuild
every image (docker compose build) at each chanhe for it to be effective.

Now we have volumes back but we do not take /node_modules in the volumes (for each docker container).
 # Please enter the commit message for your changes. Lines starting