[feat] Adding basic display if their is no activ monitoring on a probe
To create a monitoring on a probe we need to know that probe has no active monitoring on it and access a screen giving info to the user.

Changed the procesure to retrieving the user probes and their info. If there is no active monitoring it sends back 0 as active monitoring id. Also
changing import script to comment parts to have a probe without active monitoring.