[feat] Adding library files to use GY-521 (imu.py and vector3d.py) and retrieving data from GY-521 (gy521_pitch.py)
We need the pitch of the probe to calculate the density using the flottation. With the pitch of the probe and a polynomial function (2nd degree) we
can interpolate the density of the beer.

Adding the necessary library files to retrieve the gyroscopic and acceleration values from the GY-521. Using files from :
https://github.com/micropython-IMU/micropython-mpu9x50