[feat] Calling get_pitch_and_roll function into main.py. Adding function parameters and tilting parametering mode
To estimate the density using the pitch of the probe we need to use a polynomial function (2nd degree). To find the parameters of that function we
need to test the probe with multiple samples. Using a parametering mode will allow us to call an endpoint only displaying the pitch. Using it we will
find the function.

Adding the function parameters, calling the pitch and roll function into man.py and adding the probe password (forgot to add it before).
 # Please enter the commit message for your changes. Lines starting