[feat] Adding ArchiveMonitoring SQL procedure to archive a monitoring
The user needs to be able to stop a monitoring and start a new after that. Therefor we need a feature to archive the monitoring (set an end date).

Adding the ArchiveMonitoring SQL procedure. This procedure receives a monitoring id, check if it exists and then sets an end date for it.