[feat] Adding /monitoring/archive api endpoint
The user needs to be able to stop a monitoring. We already have an SQL procedure and we need an endpoint to call that procedure.

Adding /monitoring/archive endpoint. It receives a monitoring id in it's body and uses the signed cookie to check if the user is connected. It then
calls the procedure to set an end date to the monitoring.