[refactor] Using same react router route for probes with and without active monitoring
I was using two different routes (/monitoring/.... and /probe/empty/...) for displaying active monitorings on a probe and a probe
without active monitoring. Problem is, after archiving a monitoring, I need to redirect the user to the route /probe/empty/... to inform the user
there is no active monitoring on the probe. I need to have the probe id on the /monitoring/... route. Using a same route is also a better structure.

Adding ProbeData React component that either displays the Chart component (when there is an active monitoring) or another component informing the
user no active monitoring is on that probe (later replaced by monitoring creation). The condition is not placed in the menu buttons anymore (redirect
to different routes depending on if there is an active monitoring or not).

Problem I encountered : Had a conditionned rendering that did not work. I was using the monitor id as the condition value. If there is an active
monitoring the Chart component is displayed. If it was 0 (no active monitoring) the other component was displayed. Worked before with the other
structure but not here. It was because I passed monitor_id as a react router parameter. It wasn't an int as before but a string. "0" is not
considered falsy.
 # Please enter the commit message for your changes. Lines starting