[fix] Problem with the Chart component for archived monitoring. Wasn't giving the probe id.
Viewing an active monitoring had no problem. But viewing an achived monitoring returned an error. It was because I was giving for the react route the
component Chart and wasn't giving for that component the probe id or monitoring id.

Changed the react route used for archived monitoring (redirect on archived list changed). Always redirecting to a route with probe id value set
at 0. Using a probe id set at 0 means a monitoring isn't active on a probe (we COULD put it but it has no purpose). A monitoring id set at 0 means
the viewed probe has no active monitoring.