[feat] rendering archiving button only if monitoring is not archived
Once a monitoring is archived and accessible from the archived list the user should not access to the monitoring button.

Using an "is_archived" props for Chart component that is passed from PorbeData component. When equal to 1, it means the monitoring is arhived and the
button is not displayed. Also not displaying the probe id to the user.