[feat] Changing creation monitoring SQL procedure and endpoint to send back new monitoring id and redirect to data page after creation.
After creating a new monitoring on a probe we would like the user to be redirected directly to the page to see it's data. To do so we need the
monitoring id of the monitoring we just created.

Changed CreateMonitoring SQL procedure to select the id of the newly created monitoring. Also changed the endpoint and the frontend code to manage
that change and use it to redirect to the monitoring data display page.