[fix] Handling error in ProbeList component when the user is not connected.
The ProbeList component is fetching data about the user probes to create a list of probes. We retrieve a list of probes with an endpoint and use
.map() on that list. However, when the user is not connected (no session or cookie but has the isConnected value in local storage) we do not receive
a list but a message we cannot use .map() on. This creates an error.

Chaning the ProbeList and MenuDrawer component. When we have a connection error message MenuDrawer does not give the data directly to ProbeList but
gives it the false value. Inside ProbeList we check if the value received is falsey or not. If it is we display a button signaling the user is not
connected and redirecting to the user connection page. Else, acts normally.