[fix] Fixing error handling with /probe/get_all endpoint
Was using a condition where the response of the endpoint call was an error message saying the user was not connected (cookie or session). However I
could not do it like that beceause I coded the endpoint so it had a code error 400. I had to put that conndition inside the catch because an error
400 400 code is considered an error.

 # Please enter the commit message for your changes. Lines starting