[refactor] Replacing CheckUserPasswordMatch procedure with GetStoredPasswordAndUserId
With the previous SQL procedure I was checking if the password was correct in the procedure. But if I use Argon2 I can't check it on the DB. I need
to retrieve the stored password and use checkPasswordArgon2() to compare the stored password and the password given by the user.

The procedure takes one argument : the user email. It send back the user id and user password (double hashed) if the account exists.