[refactor] Addapting /user/connection endpoint and connection form to use the changed SQL procedure.
The procedure doesn't check anymore if the password is correct. It sends back the user id and the stored password.

We use checkPasswordArgon2() function to compare the stored password and the password given by the user. It returns a awnser depending on if the
password given is correc or not. Adding the frontend hash before calling the endpoint from the form.