[refactor] Change DeleteProbe SQL procedure so it checks if their is not an active monitoring on the probe.
We don't want to delete a probe if their is an active monitoring on it. We could automatically close it but I think it is better to not work like
this.

The procedure checks if theere is an active monitoring on the probe. If there is, nothing is done and the procedure sends a warning message back.