[fix +++] Fixing problems with monitoring archiving retrieve and probe list for MenuDrawer
When retrieving archived monitorings and active probes for probe list in drawer we need to have a difference between the probes retrieved. The probe
list needs the active probes and the archiving list needs all the probes related to the user to find all the archived monitorings related to the
user's probes.

I thus added an SQL procedure and api endpoint (/probe/get_active) to have a difference.

Also corrected some code in /probe/get_active to manage the case in which there are no active monitorings. I checked a condition on a SQL response
that did not exist (probes[0][0] is undefined). I need to do the same for /probe/get_all.