[fix] Correcting "homepage" value in package.json for frontend
When running the app on the server I had a problem when not on the root. When I changed the page (different URL) I had a problem with syntax (unexpected token
'<') and manifest.json not being found. My guess is the client was looking for files he should only look when he is on the root. But when using "./" as homepage
value he always thought he was on the root (maybe).