[fix] Fixing salt not working for sha256 hash on frontend
I was using the CryptoJS.SHA256() function with two arguments : the password and the salt to be used to hash everything. The
function does not use two parameters. I needed to concatene the password and the salt before passing it to the function to create
the hash. # Please enter the commit message for your changes. Lines starting