[refactor] Changing Base and ArchivedList components for testing purposes
We cannot render ArchivedList on its own in react tests. That is because ArchivedList uses a "Link" component proper to React Router. This Link component has to be
inside Router tags (BrowserRouter, MemoryRouter). So we have to render the base component to test ArchivedList behavior. ArchivedList is rendered on a certain url
route (/monitoring/archived). To do that in testig we added the possibility for the base component to have a prop allowing fixing the url route. We use <MemoryRouter>
to force that route.

 # Please enter the commit message for your changes. Lines starting