[feat] Reading tcd1304 values through DMA of RP2040
The datarate of the tcd1304 is far to high. To properly read the data it must be passed directly to the memory with a FIFO DMA. tcd1304_script.py does that. tcd1304.py
contains a class that has the same purpose.

The script and the class do not give the same results. Also, there is a problem when reading 3694 values throught DMA. An error occurs.