[refactor] Changing table structure in ArhcivedList component to run tests correctly.
I was using TableRow tags from MUI in the component. I gave them the "Link" component prop from React Router that allows a user to be redirected when
clicked on them. I had this error : "Warning: validateDOMNesting(...): <td> cannot appear as a child of <a>."

That is because the Link component created a <a> tag that becomes parent of the <td> tag created by TableRow. That is not possible.
Changed that. Less nice when looking at it but still redirects when clicking on the monitoring name.